+++ validator/control_node/requirements.txt
@@ -21,5 +21,4 @@ opentelemetry-instrumentation-system-metrics==0.47b0
 asyncclick==8.1.7.2
 rich==13.7.1
 nltk==3.9.1
-aiofiles==24.1.0
 # numpy==1.26.4
