+++ validator/utils/synthetic/synthetic_constants.py
@@ -1,4 +1,5 @@
 MAX_SEED = 100000000
+RANDOM_TEXT_QUEUE_MAX_SIZE = 500 
 
 SEED = "seed"
 TEMPERATURE = "temperature"
