+++ validator/utils/synthetic/synthetic_utils.py
@@ -1,6 +1,5 @@
 import random
 from typing import Any, Tuple
-
 from core import task_config as tcfg
 from validator.utils.redis import (
     redis_utils as rutils,
@@ -9,28 +8,26 @@ from validator.utils.redis import (
 from validator.utils.synthetic import synthetic_constants as scst
 from redis.asyncio import Redis
 from core.models import config_models as cmodels
-
+from nltk.tokenize import sent_tokenize, word_tokenize
+import numpy as np
+from time import time
+import re
 import base64
 from io import BytesIO
 import asyncio
-from contextlib import contextmanager
 import aiohttp
 import diskcache
-import aiofiles
 from PIL import Image
 import uuid
 import numpy as np
 from fiber.logging_utils import get_logger
 import random
-import os
-import nltk
-import fcntl
 import json
 from functools import lru_cache
 
 logger = get_logger(__name__)
 
-random_text_queue = asyncio.Queue(maxsize=500)
+random_text_queue = asyncio.Queue(maxsize=scst.RANDOM_TEXT_QUEUE_MAX_SIZE)
 
 @lru_cache(maxsize=None)
 def get_synth_corpus():
@@ -42,6 +39,80 @@ def get_synth_corpus():
             synth_corpus = json.load(fh)
     return synth_corpus
 
+
+def split_sentences(text):
+    fragments = sent_tokenize(text)
+    return [frag for frag in fragments if len(frag.split()) > 2]
+
+async def get_random_text_from_queue(): 
+    try:
+        if not random_text_queue.empty():
+            return await random_text_queue.get()
+        return None
+    except Exception as e:
+        logger.error(f"Error retrieving text from queue: {e}")
+        return None
+
+async def generate_text(corpus, n_words):
+    random.seed(time()%10000)
+    generated_text_parts = []
+
+    current_word_count = 0
+    categories = list(corpus.keys())
+
+    while current_word_count < n_words:
+        random.shuffle(categories)
+        # randomly select text from random categories, until we reach n_words
+        for i, category in enumerate(categories):
+            sentence = random.choice(corpus[category]).strip()
+            sentences_in_category = split_sentences(sentence)
+
+            if not sentences_in_category:
+                continue
+
+            if i > 0 and i%3 == 0:
+                sentence_part = await get_random_text_from_queue()
+            else:    
+                sentence_part = random.choice(sentences_in_category)
+
+            if not sentence_part:
+                continue
+            
+            sentence_word_count = len(word_tokenize(sentence_part))
+            if current_word_count + sentence_word_count > n_words:
+                remaining_words = n_words - current_word_count
+                truncated_part = ' '.join(sentence_part.split()[:remaining_words])
+                generated_text_parts.append(truncated_part)
+                current_word_count += remaining_words
+                break
+
+            generated_text_parts.append(sentence_part)
+            current_word_count += sentence_word_count
+
+            if current_word_count >= n_words:
+                break
+
+        if not generated_text_parts:
+            raise ValueError("Unable to generate text, problem with corpus?")
+        
+    merged_text = ' '.join(generated_text_parts).strip()
+    possible_endings = ['.', '!', '?', '...']
+
+    if merged_text and merged_text[-1] not in possible_endings:
+        if random.choice([True, False]):
+            merged_text += random.choice(possible_endings)
+
+    merged_text = re.sub(r'[^\x20-\x7E]', '', merged_text).strip()
+    return merged_text
+
+def get_random_int_from_dist(size=1, gamma_mean=1000, max_value=8000, gamma_shape=0.5, gaussian_mean=1000, gaussian_weight=0.3, gaussian_std=850):
+    gamma_scale = gamma_mean / gamma_shape
+    gamma_samples = np.random.gamma(gamma_shape, gamma_scale, size)
+    gaussian_samples = np.random.normal(gaussian_mean, gaussian_std, size)
+    combined_samples = gaussian_weight * gaussian_samples + (1 - gaussian_weight) * gamma_samples
+    combined_samples = combined_samples[combined_samples < max_value]
+
+    return combined_samples
 
 async def fetch_random_text() -> Tuple[str, int, int]:
 
@@ -72,14 +143,14 @@ async def get_save_random_text() -> None:
         try:
             queue_size = random_text_queue.qsize()
             
-            if queue_size < 500:
+            if queue_size < scst.RANDOM_TEXT_QUEUE_MAX_SIZE:
                 text, n_paragraphes, n_sentences = await fetch_random_text()                
                 await random_text_queue.put(text)
                 logger.info(f"Pushed random metaphorpsum.com text with {n_paragraphes} paragraphs, and {n_sentences} sentences to queue")
             else:
                 logger.info("Queue is full. Skipping text insertion")                
             
-            await asyncio.sleep(1)
+            await asyncio.sleep(2)
             
         except Exception as e:
             logger.error(f"Error fetching and saving synthetic data: {e} - sleeping for 60s")
